C - Functions, nested loops

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,7 +7,7 @@
  */
 void print_fib(int n)
 {
-	int f = 0, i = 1, b;
+	unsigned long f = 0, i = 1, b;
 
 	if (n < 1)
 	{
@@ -17,7 +17,7 @@ void print_fib(int n)
 	printf("%lu ", f);
 	for (b = 1; b < n; b++) 
 	{
-		int next;
+		unsigned long next;
 
 		printf("%lu ", i);
 		next = f + i;
